SupervisedTrainer/requirements.txt
-c constraints.txt
numpy==1.16.2
opencv-python==4.0.0.21
pypiwin32==223
keyboard==0.13.3
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
pip==19.0.3
protobuf==3.7.1
pyparsing==2.3.1
python-dateutil==2.8.0
pywin32==224
setuptools==40.8.0
six==1.12.0
tensorboard==1.13.1
termcolor==1.1.0
Werkzeug==0.15.1
wheel==0.33.1
certifi==2019.3.9
chardet==3.0.4
future==0.17.1
gym==0.12.1
idna==2.8
mock==2.0.0
pbr==5.1.3
Pillow==5.4.1
pyglet==1.3.2
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
tensorflow-estimator==1.13.0
urllib3==1.24.1
matplotlib==3.0.3
tensorflow-gpu==1.10.0
